Read the example Data Package

Reads the example Data Package included in frictionless. This dataset is used in examples, vignettes, and tests and contains dummy camera trap data organized in 3 Data Resources:

deployments: one local data file referenced in "path": "deployments.csv".

observations: two local data files referenced in "path": ["observations_1.tsv", "observations_2.tsv"].

media: inline data stored in data.

Usage

example_package(version = "2.0")

Arguments

version

Data Package standard version number.

Value

A Data Package object, see create_package().

Details

The example Data Package is available in two versions:

1.0: specified in the v1 specification.

2.0: specified in the v2 specification.

Examples

# Version 2 (default)
example_package()
#> A Data Package (version 2.0) with 3 resources:
#> • deployments
#> • observations
#> • media
#> Use `unclass()` to print the Data Package as a list.

# Version 1
example_package(version = "1.0")
#> A Data Package (version 1.0) with 3 resources:
#> • deployments
#> • observations
#> • media
#> Use `unclass()` to print the Data Package as a list.