Get a Data Resource

Returns a Data Resource from a Data Package, i.e. the content of one of the described resources.

get_resource(resource_name, package)

Arguments

resource_name
Name of the resource.

package
Data Package object, see read_package().

Value

List object describing a Data Resource.

Examples

# Read datapackage.json file
package <- read_package(
  system.file("extdata", "datapackage.json", package = "frictionless")
)
#> Please make sure you have the right to access data from this Data Package for your proposed use.
#> Follow applicable norms or requirements to credit the dataset and its authors.

# Get resource "observations"
resource <- get_resource("observations", package)
str(resource)
#> List of 8
#>  $ name     : chr "observations"
#>  $ path     : chr [1:2] "observations_1.csv" "observations_2.csv"
#>  $ profile  : chr "tabular-data-resource"
#>  $ title    : chr "Camera trap observations"
#>  $ format   : chr "csv"
#>  $ mediatype: chr "text/csv"
#>  $ encoding : chr "utf-8"
#>  $ schema   :List of 4
#>   ..$ fields       :List of 7
#>   .. ..$ :List of 3
#>   .. .. ..$ name       : chr "observation_id"
#>   .. .. ..$ type       : chr "string"
#>   .. .. ..$ constraints:List of 2
#>   .. .. .. ..$ required: logi TRUE
#>   .. .. .. ..$ unique  : logi TRUE
#>   .. ..$ :List of 3
#>   .. .. ..$ name       : chr "deployment_id"
#>   .. .. ..$ type       : chr "string"
#>   .. .. ..$ constraints:List of 1
#>   .. .. .. ..$ required: logi TRUE
#>   .. ..$ :List of 4
#>   .. .. ..$ name       : chr "timestamp"
#>   .. .. ..$ type       : chr "datetime"
#>   .. .. ..$ format     : chr "%Y-%m-%dT%H:%M:%S%z"
#>   .. .. ..$ constraints:List of 1
#>   .. .. .. ..$ required: logi TRUE
#>   .. ..$ :List of 3
#>   .. .. ..$ name       : chr "scientific_name"
#>   .. .. ..$ type       : chr "string"
#>   .. .. ..$ constraints:List of 1
#>   .. .. .. ..$ required: logi FALSE
#>   .. ..$ :List of 3
#>   .. .. ..$ name       : chr "count"
#>   .. .. ..$ type       : chr "integer"
#>   .. .. ..$ constraints:List of 2
#>   .. .. .. ..$ required: logi FALSE
#>   .. .. .. ..$ minimum : int 1
#>   .. ..$ :List of 3
#>   .. .. ..$ name       : chr "age"
#>   .. .. ..$ type       : chr "string"
#>   .. .. ..$ constraints:List of 2
#>   .. .. .. ..$ required: logi FALSE
#>   .. .. .. ..$ enum    : chr [1:5] "adult" "subadult" "juvenile" "offspring" ...
#>   .. ..$ :List of 3
#>   .. .. ..$ name       : chr "comments"
#>   .. .. ..$ type       : chr "string"
#>   .. .. ..$ constraints:List of 1
#>   .. .. .. ..$ required: logi FALSE
#>   ..$ missingValues: chr [1:3] "" "NA" "NaN"
#>   ..$ primaryKey   : chr "observation_id"
#>   ..$ foreignKeys  :List of 1
#>   .. ..$ :List of 2
#>   .. .. ..$ fields   : chr "deployment_id"
#>   .. .. ..$ reference:List of 2
#>   .. .. .. ..$ resource: chr "deployments"
#>   .. .. .. ..$ fields  : chr "deployment_id"