Add a Data Resource

Adds a Tabular Data Resource to a Data Package. The resource will be a Tabular Data Resource. The resource name can only contain lowercase alphanumeric characters plus ., - and _.

add_resource(package, resource_name, df, schema = NULL)

Arguments

package

List object describing a Data Package, created with read_package() or create_package().

resource_name

Name of the Data Resource.

df

A data frame.

schema

List object describing a Table Schema for that data frame. If not provided, one will be created (using create_schema()).

Value

Provided package with one additional resource.

Examples

# Load the example Data Package
package <- example_package

# List the resource names
package$resource_names
#> [1] "deployments"  "observations" "media"       

# Create a data frame
df <- data.frame(
  multimedia_id = c(
    "aed5fa71-3ed4-4284-a6ba-3550d1a4de8d",
    "da81a501-8236-4cbd-aa95-4bc4b10a05df"
  ),
  x = c(718, 748),
  y = c(860, 900)
)

# Add the data frame as a new resource to the Data Package
package <- add_resource(package, "positions", df)

# List the resource names ("positions" added)
package$resource_names
#> [1] "deployments"  "observations" "media"        "positions"