Write a Data Package to disk

Writes a Data Package and its related Data Resources to disk as a datapackage.json and CSV files. Originally included resources are left untouched, those added with add_resource() are written to file using readr::write_csv().

write_package(package, directory = ".")

Arguments

package

List object describing a Data Package, created with read_package() or create_package().

directory

Path to local directory to write files to.

Value

Provided package (invisibly).

Examples

# Load the example Data Package from disk
package <- read_package(
  system.file("extdata", "datapackage.json", package = "frictionless")
)
#> Please make sure you have the right to access data from this Data Package for your intended use.
#> Follow applicable norms or requirements to credit the dataset and its authors.

# List the resource names
package$resource_names
#> [1] "deployments"  "observations" "media"       

# Write the (unchanged) Data Package to disk
write_package(package, directory = "my_package")

# Check files
list.files("my_package")
#> [1] "datapackage.json"   "deployments.csv"    "observations_1.csv"
#> [4] "observations_2.csv"

# Clean up (don't do this if you want to keep your files)
unlink("my_package", recursive = TRUE)